--- gpg/doc.go
@@ -1,2 +1,0 @@
-// Package gpg contains implementation of terraform-provider-opengpg.
-package gpg
