+++ main.go
@@ -3,11 +3,11 @@ package main
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/plugin"
 
-	"github.com/coopnorge/terraform-provider-opengpg/gpg"
+	"github.com/coopnorge/terraform-provider-opengpg/opengpg"
 )
 
 func main() {
 	plugin.Serve(&plugin.ServeOpts{
-		ProviderFunc: gpg.Provider,
+		ProviderFunc: opengpg.Provider,
 	})
 }
