+++ opengpg/doc.go
@@ -1,0 +1,2 @@
+// Package opengpg contains implementation of terraform-provider-opengpg.
+package opengpg
