+++ opengpg/provider_test.go
@@ -1,9 +1,9 @@
-package gpg_test
+package opengpg_test
 
 import (
 	"testing"
 
-	"github.com/coopnorge/terraform-provider-opengpg/gpg"
+	"github.com/coopnorge/terraform-provider-opengpg/opengpg"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
@@ -11,15 +11,15 @@ import (
 // The factory function will be invoked for every Terraform CLI command executed
 // to create a provider server to which the CLI can reattach.
 var providerFactories = map[string]func() (*schema.Provider, error){
-	"gpg": func() (*schema.Provider, error) {
-		return gpg.Provider(), nil
+	"opengpg": func() (*schema.Provider, error) {
+		return opengpg.Provider(), nil
 	},
 }
 
 func TestProvider(t *testing.T) {
 	t.Parallel()
 
-	provider := gpg.Provider()
+	provider := opengpg.Provider()
 
 	if err := provider.InternalValidate(); err != nil {
 		t.Fatalf("validating provider internally: %v", err)
